# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StillGood App API
  description: |
    API для приложения StillGood.

    ## Роли
    - **USER** — поиск, бронирование, заказы
    - **PARTNER** — управление боксами, аналитика
    - **ADMIN** — веб-панель модерации

    ## Аутентификация
    - Access Token: 30 минут
    - Refresh Token: 14 дней
    - JWT Bearer token в заголовке Authorization
  version: 1.0.0
  contact:
    name: StillGood Development Team
    email: team@stillgood.tech
  x-team:
    - role: Tech Lead, Android Developer
      email: dmitry@stillgood.tech
    - role: Backend Developer
      email: timofey@stillgood.tech
    - role: Backend Developer
      email: artem@stillgood.tech
servers:
  - url: https://api.stillgood.tech/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
...
  /partner/analytics:
    get:
      tags: [Analytics]
      summary: Аналитика партнёра
      operationId: getPartnerAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Дата начала периода в формате YYYY-MM-DD. Если не передана или передана некорректно, используется период за последние 7 дней.
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Дата конца периода в формате YYYY-MM-DD. Если не передана или передана некорректно, используется текущая дата.
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: day
      responses:
        '200':
          description: Статистика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '403':
          description: Партнер не одобрен (PARTNER_NOT_APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Недопустимое значение groupBy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка получения аналитики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== RESTAURANTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AnalyticsResponse:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalBookings:
              type: integer
            completedOrders:
              type: integer
            cancelledOrders:
              type: integer
            expiredOrders:
              type: integer
            totalRevenue:
              type: number
            serviceFee:
              type: number
            netPayout:
              type: number
            conversionRate:
              type: number
            cancelRate:
              type: number
            categoryBreakdown:
              type: array
              items:
                $ref: '#/components/schemas/CategoryStatDto'
        periods:
          type: array
          items:
            $ref: '#/components/schemas/DailyAnalyticsDto'
    CategoryStatDto:
      type: object
      properties:
        name:
          type: string
        totalRevenue:
          type: number
    DailyAnalyticsDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        category_name:
          type: string
        total_bookings:
          type: integer
        completed_orders:
          type: integer
        cancelled_orders:
          type: integer
        expired_orders:
          type: integer
        gross_revenue:
          type: number
        service_fee:
          type: number
        net_payout:
          type: number
        created_at:
          type: string
          format: date-time

    # ==================== COMMON ====================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Человекочитаемое описание
...